Split header and explicitly instantiate template functions.
